Finished MaintainabilityIndex Integration Tests
Split integration test into two.
CyclomaticComplexity and ExecutableStatementCount are spys so that I can test getMaintainabilityIndex better.

CommentCounter is a spy so that I can have it act as a stub in the first test and a real class in the second.

First test is testing integration between MaintainabilityIndex and HalsteadMetrics while fully mocking CommentCounter.

Secondtest is testing integration between MaintainabilityIndex, HalsteadMetrics, and CommentCounter.